refactor(ReservationTime): 클래스 명 변경
