+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -6,20 +6,20 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.web.bind.annotation.*;
-import roomescape.domain.Time;
-import roomescape.dto.TimeRequestDto;
-import roomescape.dto.TimeResponseDto;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.ReservationTimeRequestDto;
+import roomescape.dto.ReservationTimeResponseDto;
 
 import javax.sql.DataSource;
 import java.util.List;
 
 @RestController
 @RequestMapping("/times")
-public class TimeController {
+public class ReservationTimeController {
     private final JdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert simpleJdbcInsert;
 
-    public TimeController(final JdbcTemplate jdbcTemplate, final DataSource dataSource) {
+    public ReservationTimeController(final JdbcTemplate jdbcTemplate, final DataSource dataSource) {
         this.jdbcTemplate = jdbcTemplate;
         this.simpleJdbcInsert = new SimpleJdbcInsert(dataSource)
                 .withTableName("RESERVATION_TIME")
@@ -27,19 +27,19 @@ public class TimeController {
     }
 
     @PostMapping
-    public TimeResponseDto create(@RequestBody final TimeRequestDto request) {
+    public ReservationTimeResponseDto create(@RequestBody final ReservationTimeRequestDto request) {
         SqlParameterSource params = new MapSqlParameterSource()
                 .addValue("start_at", request.getStartAt());
         Long id = simpleJdbcInsert.executeAndReturnKey(params).longValue();
-        return new TimeResponseDto(id, request.getStartAt());
+        return new ReservationTimeResponseDto(id, request.getStartAt());
     }
 
     @GetMapping
-    public List<TimeResponseDto> retrieveTimes() {
-        List<Time> timeList = jdbcTemplate.query("select * from reservation_time", timeRowMapper);
-        return timeList.stream()
-                       .map(TimeResponseDto::from)
-                       .toList();
+    public List<ReservationTimeResponseDto> retrieveTimes() {
+        List<ReservationTime> reservationTimeList = jdbcTemplate.query("select * from reservation_time", timeRowMapper);
+        return reservationTimeList.stream()
+                                  .map(ReservationTimeResponseDto::from)
+                                  .toList();
     }
 
     @DeleteMapping("/{id}")
@@ -48,11 +48,11 @@ public class TimeController {
         jdbcTemplate.update(sql, Long.valueOf(id));
     }
 
-    private final RowMapper<Time> timeRowMapper = (resultSet, rowNum) -> {
-        Time time = new Time(
+    private final RowMapper<ReservationTime> timeRowMapper = (resultSet, rowNum) -> {
+        ReservationTime reservationTime = new ReservationTime(
                 resultSet.getLong("id"),
                 resultSet.getString("start_at")
         );
-        return time;
+        return reservationTime;
     };
 }
