+++ src/main/java/roomescape/dto/ReservationTimeRequestDto.java
@@ -3,11 +3,11 @@ package roomescape.dto;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class TimeRequestDto {
+public class ReservationTimeRequestDto {
     private final String startAt;
 
     @JsonCreator
-    public TimeRequestDto(@JsonProperty("startAt") final String startAt) {
+    public ReservationTimeRequestDto(@JsonProperty("startAt") final String startAt) {
         this.startAt = startAt;
     }
